finish book 3

+++ preview.py
@@ -1,3 +1,4 @@
+import time
 import os
 import matplotlib
 import matplotlib.pyplot as plt
@@ -6,13 +7,14 @@ import subprocess
 import sys
 
 filename = sys.argv[1]
+print("ctrl-c to quit at any time")
 # set recursion limit much higher
 sys.setrecursionlimit(10**6)
-
+done = False
 # taken from https://stackoverflow.com/a/45734500
 def mypause(interval):
     backend = plt.rcParams['backend']
-    if backend in matplotlib.rcsetup.interactive_bk:
+    if backend in matplotlib.backends.backend_registry.list_builtin(matplotlib.backends.BackendFilter.INTERACTIVE):
         figManager = matplotlib._pylab_helpers.Gcf.get_active()
         if figManager is not None:
             canvas = figManager.canvas
@@ -24,8 +26,16 @@ def mypause(interval):
 
 
 def _update():
+    global done
     result = subprocess.run(["python", "multiprocess.py", filename], capture_output=True, text=True)
-    return result.stdout.split('\n')[-2].split(' ')[-1]
+    try:
+        fn = result.stdout.split('\n')[-2].split(' ')[-1]
+    except IndexError:
+        done = True
+        time.sleep(0.5)
+        print("Done")
+        return f"out/{filename}.png"
+    return fn
 
 
 def update_image(ax, fig):
@@ -34,7 +44,14 @@ def update_image(ax, fig):
     ax.clear()
     ax.imshow(img)
     fig.canvas.draw()
-    mypause(3)
+    if done:
+        return
+    try:
+        mypause(3)
+    except KeyboardInterrupt:
+        print("\nQuitting...")
+        sys.exit(1)
+
     os.remove(preview_filename)
     update_image(ax, fig)  # Recursive update
 
@@ -50,7 +67,14 @@ ax.imshow(img)
 plt.axis('off')  # Turn off axes
 
 # Start the recursive image update
-update_image(ax, fig)
+if not done:
+    update_image(ax, fig)
 
+plt.ioff()
+print(f"Final file in out/{filename}.png")
 # Display the plot
-plt.show()
+try:
+    plt.show()
+except KeyboardInterrupt:
+    print("\nQuitting...")
+    sys.exit(1)
